fixed tag Bug + refactoring
